requirements.txt
Django==1.11.6
djangorestframework==3.7.0
django-filter==1.1.0
django-widget-tweaks=1.4.2
